ada: VAST found bug: Missing Parent in annotate aspect
In case of an Annotate aspect of the form "Annotate => Expr",
where Expr is an identifier (as opposed to an aggregate),
the Parent field of the N_Identifier node for Expr was
destroyed. This patch changes the code that turns the aspect
into a pragma, so that it no longer has that bug.

The problem was in "New_List (Expr)"; which sets the Parent of
Expr to Empty. But Expr is still part of the tree of the aspect,
so it should have a proper Parent; we can't just stick it in a
temporary list.

The new algorithm constructs the pragma arguments without disturbing
the tree of the aspect.

This is the last known case of missing Parent fields, so we can
now enable the VAST check that detected this bug.

gcc/ada/ChangeLog:

	* sem_ch13.adb (Aspect_Annotate): Avoid disturbing the tree of the
	aspect.
	* vast.adb: Enable Check_Parent_Present.
	* exp_ch6.adb (Validate_Subprogram_Calls): Minor reformatting.