Match: Simplify (T1)(a bit_op (T2)b) to (T1)a bit_op (T1)b
During the match pattern of SAT_U_MUL form 7, we found there is
a pattern like below:

(nop_convert)(a bit_op (convert b))

which result in the pattern match of SAT_U_MUL complicated and
unintuitive.  According to the suggestion of Richard, we would
like to simply it to blew:

(convert a) bit_op (convert b)

which is more friendly for reading and bit_op.  There are three
bit_op here, aka bit_ior, bit_and and bit_xor.

gcc/ChangeLog:

	* match.pd: Add simplfy to fold outer convert of bit_op
	to inner captures.

Signed-off-by: Pan Li <pan2.li@intel.com>