forwprop: Allow nop conversions for vector constructor.
I observed a vect-construct forwprop opportunity in x264 that we
could handle when checking for a nop conversion instead of a useless
conversion.  IMHO a nop-conversion check is sufficient as we're only
dealing with permutations in simplify_vector_constructor.

This patch replaces uses of useless_type_conversion_p with
tree_nop_conversion_p in simplify_vector_constructor.

It was bootstrapped and regtested on x86 and power10, regtested on
aarch64 and riscv64.

There is a single scan-test failure on power
(gcc.target/powerpc/builtins-1.c).  The code actually looks better so
I took the liberty of adjusting the test expectation.

gcc/ChangeLog:

	* tree-ssa-forwprop.cc (simplify_vector_constructor):
	Allow nop conversions.

gcc/testsuite/ChangeLog:

	* gcc.target/powerpc/builtins-1.c: Adjust test expectation.
	* gcc.target/riscv/rvv/autovec/pr118019-3.c: New test.