aarch64: Cache the PCS value for a function
As aarch64_function_arg_regno_p is a very hot function called many times for a
function it should not call fndecl_abi every time as it is expensive and
unecessasary.

This caches the result and in doing so fixes a regression in compile time
introduced by r16-5076-g7197d8062fddc2.

gcc/ChangeLog:

	* config/aarch64/aarch64.cc (aarch64_fndecl_abi): New function.
	(aarch64_function_abi): New function.
	(aarch64_function_arg_regno_p): Update to use aarch64_fndecl_abi.
	(aarch64_output_mi_thunk): Likewise.
	(aarch64_is_variant_pcs): Likewise.
	(aarch64_set_current_function): Update to initialize pcs value.
	* config/aarch64/aarch64.h (enum arm_pcs): Move location earlier in
	file.
	(machine_function) Add pcs value.