c++: Handle absolute path for CMI output directory [PR122677]
When trying to create each directory component of an absolute path, the
first call to mkdir uses a path of "" which gives an ENOENT error. This
causes the create_dirs function to return without creating anything.

This commit skips past the leading slashes of an absolute path, so that
the first call to mkdir is for an actual directory name, not an empty
string.

gcc/cp/ChangeLog:

	PR c++/122677
	* module.cc (create_dirs): Skip past any leading slashes.

(cherry picked from commit c1cf465bdbe587b1ccd8feb05b9b3878163708fe)