doc, darwin: Clean up Darwin options and documentation [PR122243]
The Darwin target options documentation was a bit of a mess, with several
undocumented options, some that were listed in the option summary or mentioned
in discussion of other options but not actually documented, and a large
number of options listed in darwin.opt as being obsolete.

I've undocumented all the obsolete options to streamline things, plus a few
others that seem to have been intentially undocumented or supplanted by
other options.  For the others that were probably supposed to documented,
I did my best to guess what they're for by reading the code or just
copying the documentation string in the .opt file, but it's certainly
possible I screwed some up.

gcc/ChangeLog
	PR other/122243
	* config/darwin.opt (findirect-virtual-calls): Mark as Undocumented.
	(fterminated-vtables): Likewise.
	(multi_module): Likewise.
	(multiply_defined): Likewise.
	(multiply_defined_unused): Likewise.
	(no_dead_strip_inits_and_terms): Likewise.
	(noprefixbinding): Likewise.
	(nomultidefs): Likewise.
	(noprebind): Likewise.
	(noseglinkedit): Likewise.
	(ObjC, ObjC++): Add documentation strings.
	(object): Mark as Undocumented.
	(prebind): Likewise.
	(prebind_all_twolevel_modules): Likewise.
	(private_bundle): Likewise.
	(sectobjectsymbols): Likewise.
	(sectorder): Likewise.
	(seg_addr_table_filename): Likewise.
	(segcreate): Likewise.
	(seglinkedit): Likewise.
	(single_module): Likewise.
	(X): Likewise.
	(y): Likewise.
	(Mach): Likewise.
	* doc/invoke.texi (Option Summary) <Darwin Options>: Improve
	alphabetization of the list.  Remove obsolete/undocumented
	options and add missing entries.
	(Darwin Options): Add documentation for -arch, -dependency-file,
	-fapple-kext, -matt-stubs, -fconstant-cfstrings, -mdynamic-no-pic,
	-asm_macosx_version_min, -msymbol-stubs, -mtarget-linker,
	-ObjC, -ObjC++, -Wnonportable-cfstrings.  Update the list
	of options passed to the linker to remove obsolete options
	and add missing ones; also move the block of @opindex entries
	before the list items instead of putting it in the middle.