ada: Fix spurious exceptions with iterated aggregates
When an array aggregate has an iterated component association over a
range that we know is empty, we don't create a loop during expansion but
we still analyze the expression of the component association in a
unusual context.

Before this patch, this analysis could incorrectly insert actions in an
enclosing scope. This patch fixes it by only doing preanalysis of the
expression in that case.

gcc/ada/ChangeLog:

	* exp_aggr.adb (Gen_Loop): Only preanalyze expressions we know won't
	evaluated.