c++: Fix error recovery ICE in tsubst_baselink [PR120876]
The following testcase ICEs since r12-6080.  The problem is that
lookup_fnfields can return NULL_TREE on failure, but the maybe_incomplete
handling was added before the if (!baselink) handling and assumes that
baselink is non-NULL (and BASELINK).

The following patch reorders the if (maybe_incomplete) handling with
if (!baselink).

2025-11-20  Jakub Jelinek  <jakub@redhat.com>

	PR c++/120876
	* pt.cc (tsubst_baselink): Move maybe_incomplete handling after
	!baselink handling.

	* g++.dg/parse/crash81.C: New test.