gcc/doc: Fix usages of @itemize with named items in Regs and Memory
The old usage lead to makeinfo emitting text like "pseudo registers This
is the most common case. Most subregs have pseudo regs as their first
operand." when, clearly, "pseudo registers" was intended to be the name
of an item, and the rest of the paragraph a description.  This is
because @itemize was used, which does not support "naming" items.

Now, the paragraph used as an example above looks like this:

  pseudo registers
       This is the most common case.  Most ‘subreg’s have pseudo
       ‘reg’s as their first operand.

gcc/ChangeLog:

	* doc/rtl.texi (Regs and Memory): Use @table instead of @itemize
	for lists with named items.