middle-end: add target hook for isel
This adds a new target hook to gimple-sel to allow targets to
do target specific massaging of the gimple IL to prepare for
expand.

Tejas will be sending up part 2 of this soon to help convert
SVE packed boolean VEC_COND_EXPR into something that we can
handle more efficiently if expanded in a different order.

We also want to use this to e.g. for Adv. SIMD prefer avoiding
!= vector compare expressions because the ISA doesn't have
this instruction and so we expand to == + ~ but changing the
expression from a MIN to MAX only for VECTOR_BOOLEAN_TYPE_P
and flipping the operands we can expand more efficient.

These are the kind of things we want to use the hook for,
not generic changes that apply to all target.

gcc/ChangeLog:

	* target.def (instruction_selection): New.
	* doc/tm.texi.in: Document it.
	* doc/tm.texi: Regenerate
	* gimple-isel.cc (pass_gimple_isel::execute): Use it.
	* targhooks.cc (default_instruction_selection): New.
	* targhooks.h (default_instruction_selection): New.