ipa: Rewrite callback_edge_useful_p heuristic.
This function is used to determine whether a callback edge should be
kept or not.  It was supposed to capture the idea that a callback edge
has been redirected at some point, however, it only considered
redirecting to some clone.  However, an edge may not always be
redirected to a clone.  For example, common function bodies produced by
icf are not clones.  This version of this function should cover all
cases, at least for the time being.

gcc/ChangeLog:

	PR ipa/122768
	* attr-callback.cc (callback_edge_useful_p): Rewrite the
	heuristic, now consider clones as well as icf bodies.

Signed-off-by: Josef Melcr <josef.melcr@suse.com>