phiprop: Allow non-trapping loads to be proped back into the loop
First the testcase for PR60183 is no longer testing that we don't
prop into the loops for possible trapping cases. This adds phiprop-4.c
that tests that.
Second we can prop back into loops if we know the load will not trap.
This adds that check. phiprop-3.c tests that.

Bootstrapped and tested on x86_64-linux-gnu.

	PR tree-optimization/60183

gcc/ChangeLog:

	* tree-ssa-phiprop.cc (propagate_with_phi): Allow
	known non-trapping loads to happen back into the
	loop.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-ssa/phiprop-3.c: New test.
	* gcc.dg/tree-ssa/phiprop-4.c: New test.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>